+++ docs/flutter_service_worker.js
@@ -316,9 +316,9 @@ const RESOURCES = {
 "favicon.png": "5dcef449791fa27946b3d35ad8803796",
 "icons/Icon-192.png": "ac9a721a12bbc803b44f645561ecb1e1",
 "icons/Icon-512.png": "96e752610906ba2a93c65f8abe1645f1",
-"index.html": "9b702896e9840c84a6d3fb26571dd237",
-"/": "9b702896e9840c84a6d3fb26571dd237",
-"main.dart.js": "c0c106c82743d88d50ca4c0e82d1998d",
+"index.html": "4f9bfbde8f3997cfab96e8268aee2f38",
+"/": "4f9bfbde8f3997cfab96e8268aee2f38",
+"main.dart.js": "6affb379479238cd9a3299870f4ce3c0",
 "manifest.json": "9e7b34fd7c291ca2523123ef20497930",
 "version.json": "4b6db237b3514a88107a422469adfb0f"
 };