+++ lib/phone_form_field.dart
@@ -1,6 +1,7 @@
 library phone_number_input;
 
 export 'src/phone_form_field.dart';
+export 'src/selector_config.dart';
 export 'src/country_selector.dart';
 export 'src/flag_dial_code_chip.dart';
 export 'src/localization/phone_field_localization.dart';